Spotify Login Page 1st Version
The first version of the Spotify login page. This might change later on, but because login pages are so simple, I doubt it will change much, if at all. There is also a comment next to the login page explaining what will happen if the user logs in to Spotify and how it will work with the website.